backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
pydantic==2.5.0
python-multipart==0.0.6
prometheus-client==0.19.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
alembic==1.13.1
python-dotenv==1.0.0
# Optional ML deps removed for local dev speed
# torch and numpy are optional and guarded in code
# Auth dependencies
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
# Email and notifications
sendgrid==6.10.0
apscheduler==3.10.4
requests==2.31.0
# Additional utilities
email-validator==2.1.0
